+++ ext/GRAPEOptimExt.jl
@@ -4,6 +4,7 @@ import Optim
 using GRAPE: GrapeWrk, update_result!
 import GRAPE: run_optimizer, step_width, search_direction
 
+
 function run_optimizer(
     optimizer::Optim.AbstractOptimizer,
     wrk,
