+++ src/GRAPE.jl
@@ -3,6 +3,6 @@ module GRAPE
 include("workspace.jl")
 include("result.jl")
 include("optimize.jl")
-include("backend_lbfgsb.jl")  # other backends are package extensions
+include(joinpath("..", "ext", "GRAPELBFGSBExt.jl"))
 
 end
